Convert SCRIPT_RUNTIME_FEATURES enum to bitflags

+++ examples/interop.rs
@@ -134,8 +134,8 @@ impl sciter::EventHandler for EventHandler {
 
 fn check_options() {
 	sciter::set_options(sciter::RuntimeOptions::ScriptFeatures(
-		sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_SYSINFO as u8		// Enables `Sciter.machineName()`
-		| sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_FILE_IO as u8	// Enables opening file dialog (`view.selectFile()`)
+		sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_SYSINFO		// Enables `Sciter.machineName()`
+		| sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_FILE_IO	// Enables opening file dialog (`view.selectFile()`)
 	)).ok();
 
 	for arg in std::env::args() {
