+++ examples/minimal.rs
@@ -13,8 +13,8 @@ fn main() {
 
 	// Step 2: Enable the features we need in our tiscript code.
 	sciter::set_options(sciter::RuntimeOptions::ScriptFeatures(
-		sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_SYSINFO as u8		// Enables `Sciter.machineName()`.  Required for opening file dialog (`view.selectFile()`)
-		| sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_FILE_IO as u8	// Enables opening file dialog (`view.selectFile()`)
+		sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_SYSINFO		// Enables `Sciter.machineName()`
+		| sciter::SCRIPT_RUNTIME_FEATURES::ALLOW_FILE_IO	// Enables opening file dialog (`view.selectFile()`)
 		)).unwrap();
 
 	// Enable debug mode for all windows, so that we can inspect them via Inspector.
